Fix resource leak, restructure code to fix a few bugs and fatal logs
Fixed an issue with the http.HandlerFunc's handling of the request.Body.
The request.Body comment in the go net/http package is misleading - it implies
that the webserver handles the request.Body.Close() call. This is true,
but only when timeouts are set. When timeouts are not configured for the
Server, if request.Body.Close() is not called, then the body will not be
closed, and this is a resource leak. Cautious http.HandlerFunc's should
always read the body, check for an error, and then close it. In that
order.

Updated the dockerfile to use the pre-built ubi9 go toolset, and added
proper caches for go mod and the go toolchain to speed up builds. The
docker compose file has also been updated to build whenever
docker-compose up is executed. This made local dev much easier.

The k6 tests have been copied in from the quarkus repo for easier
execution, this saved me having to write proper test given the POC
nature of this work.

The go modules have been tidied up a little, including updated the
versions of dependencies.

The jwt package has been renamed to downstream, and represents the
client for the downstream api, also called "testserver". The postgres
client also has some updates, mostly to do with the pgx versio bump.